Move files to new directory
