+++ Src/lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
-import 'package:ossproj_comfyride/Login_Screen.dart';
-import 'package:ossproj_comfyride/ImageProviderNotifier.dart';
+import 'package:ossproj_comfyride/screen/Login_Screen.dart';
+import 'package:ossproj_comfyride/provider/ImageProviderNotifier.dart';
 import 'package:provider/provider.dart';
 
 Future<void> main() async {
