+++ Src/lib/screen/Style_Recommendation.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:ossproj_comfyride/ImageProviderNotifier.dart';
+import 'package:ossproj_comfyride/provider/ImageProviderNotifier.dart';
 import 'package:provider/provider.dart';
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
